Upgrade test environment to flake8 3.7.9
Due to the way that flake8 now registers its options, the approach I was
using in the TestOptions class no longer works. The code results in an
OptionConflictError as soon as Application.initialize() is called.

I haven't been able to track down what needs to be done in order for
those tests to work again so I'm going to drop them for now. The
additional coverage was nice, but I think we can safely live without it,
and the behavior of the options themselves is still tested by the main
suite.