Handle keyword arguments in assert method calls
Although it's less common to call assert methods using keyword
arguments, it's something we need to support.

This changes introduces an args() utility function that progressively
yields an ast.Call node's positional and keyword arguments. We use this
function within our checks to always consider all of a call's arguments
in order.

Fixes #7