Detect and warn on assertTrue misuse
A common error is to use the two argument form of `assertTrue()` when
one intends to use `assertEqual()`. This leads to meaningless tests that
(almost) always pass.

For example:

```
self.assertTrue(might_be_foo, "foo")
```

instead of:

```
self.assertEqual(might_be_foo, "foo")
```

The second argument is the error message on failure, and is better
specified as the `msg=` kerword argument.

This was inspired by https://richardtier.com/2022/02/16/3-of-666-python-codebases-we-checked-had-silently-failing-unit-tests-and-we-fixed-them-all%ef%bf%bc/